перенесено в пакеты

+++ src/lesson6/Employee.java
@@ -6,11 +6,11 @@ public class Employee {
     private String position;
     private String email;
     private String phone;
-    private double salary;
+    private int salary;
     private int age;
 
     // Конструктор класса
-    public Employee(String fullName, String position, String email, String phone, double salary, int age) {
+    public Employee(String fullName, String position, String email, String phone, int salary, int age) {
         this.fullName = fullName;
         this.position = position;
         this.email = email;
